% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_data.R
\name{backup_data}
\alias{backup_data}
\title{Backup Data}
\usage{
backup_data(
  directory = "",
  limit = "1000000",
  url = "https://core01.respndmi.app:7473",
  user = "mp_read_user",
  password = ""
)
}
\arguments{
\item{directory}{Directory to save backup (full path)}

\item{limit}{Maximum number of cases to fetch}

\item{url}{URL for neo4j server}

\item{user}{URL Username for neo4j server}

\item{password}{URL Password for neo4j server}
}
\value{
data.frame Person-to-person referral data
}
\description{
Backup Data
}
