% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gender_variable.R
\name{calculate_gender_variable}
\alias{calculate_gender_variable}
\title{Create Gender Variable}
\usage{
calculate_gender_variable(data)
}
\arguments{
\item{data}{Person data (data.frame)}
}
\value{
data.frame Person data with genderId variable
}
\description{
Create Gender Variable
}
