% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_race_variable.R
\name{calculate_race_variable}
\alias{calculate_race_variable}
\title{Create Race Variable}
\usage{
calculate_race_variable(data)
}
\arguments{
\item{data}{Person data (data.frame)}
}
\value{
data.frame Updated dataset with race variable
}
\description{
Create Race Variable
}
