% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_symptomcount_variable.R
\name{calculate_symptomcount_variable}
\alias{calculate_symptomcount_variable}
\title{Calculate symptom count variable}
\usage{
calculate_symptomcount_variable()
}
\description{
Calculate symptom count variable
}
