% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_tract_sf_obj.R
\docType{data}
\name{census_tract_sf_obj}
\alias{census_tract_sf_obj}
\title{Census Tract SF Object}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 7569 rows and 4 columns.
}
\usage{
census_tract_sf_obj
}
\value{
A sf object with columns: \code{census_tract} (chr), \code{county} (chr), and \code{state} (chr)
}
\description{
A simple features (sf) object of all census tracts in the United States
}
\keyword{datasets}
