% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_census_to_neighborhood.R
\name{convert_census_to_neighborhood}
\alias{convert_census_to_neighborhood}
\title{Concert census tract to neighborhood in New York City}
\usage{
convert_census_to_neighborhood(census_tract)
}
\arguments{
\item{census_tract}{}
}
\description{
Concert census tract to neighborhood in New York City
}
