% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivetable.R
\name{descriptivetable}
\alias{descriptivetable}
\title{Descriptive Table}
\usage{
descriptivetable(dat, dig)
}
\arguments{
\item{dat}{Data set (data.frame)}

\item{dig}{How many digits in the decimal place (integer)}
}
\value{
data.frame What is this going to return
}
\description{
This function helps to generte tables. It returns an object somewhere between a kable and a data.frame.
}
