% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_drive.R
\name{distance.drive}
\alias{distance.drive}
\title{Drive Distance}
\usage{
distance.drive(census_tract_from, census_tract_to, key = "", quiet = T)
}
\arguments{
\item{census_tract_from}{Origin census tract}

\item{census_tract_to}{Destination census tract}
}
\value{
vector of distances
}
\description{
Returns the shortest driving distance between a pair of census tracts
}
