% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_subway.R
\name{distance.subway}
\alias{distance.subway}
\title{Subway Distance}
\usage{
distance.subway(census_tract_from, census_tract_to, key = "")
}
\arguments{
\item{census_tract_from}{Origin census tract}

\item{census_tract_to}{Destination census tract}
}
\value{
vector of distances
}
\description{
Returns the shortest subway distance between a pair of census tracts
}
