% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_walk.R
\name{distance.walk}
\alias{distance.walk}
\title{Walking Distance}
\usage{
distance.walk(census_tract_from, census_tract_to, key = "", quiet = T)
}
\arguments{
\item{census_tract_from}{Origin census tract}

\item{census_tract_to}{Destination census tract}

\item{key}{Google Maps API Key}
}
\value{
vector of distances
}
\description{
Returns the shortest walking distance between a pair of census tracts
}
