% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_table_1.R
\name{draw_table_1}
\alias{draw_table_1}
\title{Draw Table}
\usage{
draw_table_1(data_table, list_labels)
}
\arguments{
\item{data_table}{Data table to be drawn as a table_1 object}

\item{list_labels}{List of labels}
}
\value{
kable Table
}
\description{
Draw Table
}
