% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_person_data.R
\name{get_person_data}
\alias{get_person_data}
\title{Get Person Data}
\usage{
get_person_data(
  url = "https://core01.respndmi.app:7473",
  user = "mp_read_user",
  password = "L8upDYa8MrJ6Cl2AFfe0"
)
}
\arguments{
\item{url}{URL for neo4j server}

\item{user}{URL Username for neo4j server}

\item{password}{URL Passwortd for neo4j server}
}
\value{
data.frame Person Level Data
}
\description{
Get Person Data
}
