% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_referral_data.R
\name{get_referral_data}
\alias{get_referral_data}
\title{Get Referral Data}
\usage{
get_referral_data(
  limit = 1,
  url = "https://core01.respndmi.app:7473",
  user = "mp_read_user",
  password = ""
)
}
\arguments{
\item{limit}{Maximum number of cases to fetch}

\item{url}{URL for neo4j server}

\item{user}{URL Username for neo4j server}

\item{password}{URL Password for neo4j server}
}
\value{
data.frame Person-to-person referral data
}
\description{
Get Referral Data
}
