% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_maps_fetch.R
\name{google.maps.fetch}
\alias{google.maps.fetch}
\title{Title}
\usage{
google.maps.fetch(
  census_tract_from = sample(census_tract_centroids$GEOID, 20),
  census_tract_to = sample(census_tract_centroids$GEOID, 20),
  mode_of_transport = "subway",
  parameter = "distance"
)
}
\arguments{
\item{parameter}{}
}
\description{
Title
}
