% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_1_datatable.R
\name{make_table_1_datatable}
\alias{make_table_1_datatable}
\title{Make Table 1 Data Table}
\usage{
make_table_1_datatable(
  bipartite_graph_sim,
  by = groupSex,
  variables = c(age, genderId),
  person_data = TRUE
)
}
\arguments{
\item{bipartite_graph_sim}{Simulated (re-sampled) bipartite graph}

\item{by}{Stratifying variable}

\item{variables}{Vector of analysis variables}

\item{person_data}{Logical variable indicating whether person or place data are to be used}
}
\value{
Tibble
}
\description{
Make Table 1 Data Table
}
\examples{
make_table_1_datatable(bipartite_graph_sim, by = groupSex, variables = c(age, genderId), person_data = TRUE)

}
