% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_combined.R
\name{plot_data_combined}
\alias{plot_data_combined}
\title{Plot Data Matrices Combined}
\usage{
plot_data_combined(
  data,
  input_variable,
  activity_setting_input,
  activity_setting_inputs,
  spacing = 5,
  list_labels
)
}
\arguments{
\item{data}{data}

\item{input_variable}{which}

\item{activity_setting_input}{which}

\item{activity_setting_inputs}{which}

\item{spacing}{5}

\item{list_labels}{labels}
}
\value{
ggplot object
}
\description{
Plot Data Matrices Combined
}
