% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_matrices_large.R
\name{plot_data_matrices_large}
\alias{plot_data_matrices_large}
\title{Plot Data Matrices Large}
\usage{
plot_data_matrices_large(
  data,
  input_variable,
  activity_setting_inputs,
  spacing = 5,
  list_labels
)
}
\arguments{
\item{data}{data}

\item{input_variable}{Variable}

\item{activity_setting_inputs}{c("home", "social", "sexual")}

\item{spacing}{spacing}

\item{list_labels}{labels}
}
\value{
ggplot Object
}
\description{
Plot Data Matrices Large
}
