% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_graph.R
\name{plot_map_graph}
\alias{plot_map_graph}
\title{Plot Network Graph}
\usage{
plot_map_graph(
  graph_type = "community",
  weight_threshold = 1,
  degree_threshold = 10,
  sex_only = FALSE,
  map_within_neighborhood = NULL,
  map_within_community = NULL,
  map_within_borough = NULL,
  mapping_basegraph = NULL,
  mapping_spatial_units = NULL,
  mapping_lines = ggplot2::aes(alpha = weight),
  mapping_points = ggplot2::aes(size = degree),
  options_basegraph = list(data = mpxnyc::community_sf_obj, alpha = 1, size = 0.1, color
    = "white", fill = "#009BE8"),
  options_spatial_units = list(alpha = 0.4, size = 0.1),
  options_lines = list(color = "#41BFFF", size = 0.1),
  options_points = list(color = "#41BFFF", size = 0.5),
  places_data
)
}
\arguments{
\item{graph_type}{Spatial Partitioning. Default "community". Other options are "censustract", "neighborhood", and "borough}

\item{weight_threshold}{Display parameter (numeric). Default 1.}

\item{degree_threshold}{Display parameter (numeric). Default 10.}

\item{sex_only}{Restrict to places where participant had sexual contact}

\item{map_within_neighborhood}{Restrict to within-neighborhood connections. Default NULL. Logical.}

\item{map_within_community}{Restrict to within-community connections. Default NULL. Logical.}

\item{map_within_borough}{Restrict to within-borough connections. Default NULL. Logical.}

\item{mapping_basegraph}{Accepts ggplot2::aes() object}

\item{mapping_spatial_units}{Accepts ggplot2::aes() object}

\item{mapping_lines}{Accepts ggplot2::aes() object}

\item{mapping_points}{Accepts ggplot2::aes() object}

\item{options_basegraph}{Accepts geom_sf options}

\item{options_spatial_units}{Accepts geom_sf options}

\item{options_lines}{Accepts geom_sf options}

\item{options_points}{Accepts geom_sf options}

\item{places_data}{Place dataset}
}
\value{
ggplot object
}
\description{
Plot Network Graph
}
