% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_radar_graphic.R
\name{plot_radar_demographics_by_stratum}
\alias{plot_radar_demographics_by_stratum}
\title{Demographics by Stratum Radar Plot}
\usage{
plot_radar_demographics_by_stratum(data, list_labels)
}
\arguments{
\item{data}{plot data}

\item{list_labels}{labels}
}
\value{
ggplot object
}
\description{
Demographics by Stratum Radar Plot
}
