% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_radar_graphic.R
\name{plot_radar_graphic}
\alias{plot_radar_graphic}
\title{Plot Radar Graph}
\usage{
plot_radar_graphic(data, list_labels)
}
\arguments{
\item{data}{plot data}

\item{list_labels}{labels}
}
\value{
ggplot object
}
\description{
Plot Radar Graph
}
