% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_social_interaction_matrices_inset.R
\name{plot_social_interaction_matrices_inset}
\alias{plot_social_interaction_matrices_inset}
\title{Plot Data Matrices Large}
\usage{
plot_social_interaction_matrices_inset(
  data_contact_matrices,
  input_variable,
  activity_setting_inputs,
  spacing = 5,
  list_labels
)
}
\arguments{
\item{input_variable}{Variable}

\item{activity_setting_inputs}{c("home", "social", "sexual")}

\item{spacing}{spacing}

\item{list_labels}{labels}

\item{data}{data}
}
\value{
ggplot Object
}
\description{
Plot Data Matrices Large
}
