% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_social_interaction_matrices.R
\name{plot_social_interaction_matrix}
\alias{plot_social_interaction_matrix}
\title{Plot Social Interaction Matrices}
\usage{
plot_social_interaction_matrix(
  data_contact_matrices,
  input_variable,
  activity_setting_input,
  activity_setting_inputs,
  spacing = 5,
  list_labels
)
}
\arguments{
\item{input_variable}{which}

\item{activity_setting_input}{which}

\item{activity_setting_inputs}{which}

\item{spacing}{5}

\item{list_labels}{labels}

\item{data}{data}
}
\value{
ggplot object
}
\description{
Plot Social Interaction Matrices
}
