% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subway_distance.R
\name{subway_distance}
\alias{subway_distance}
\title{Subway Distance}
\usage{
subway_distance(census_tract_a, census_tract_b)
}
\arguments{
\item{census_tract_from}{Origin census tract}

\item{census_tract_to}{Destination census tract}
}
\value{
vector of distances
}
\description{
Returns the shortest subway distance between a pair of census tracts
}
