% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_distance.R
\name{travel_distance}
\alias{travel_distance}
\title{Travel Distance}
\usage{
travel_distance(census_tract_from, census_tract_to, mode_of_transport, ...)
}
\arguments{
\item{census_tract_from}{Origin Census Tract (string)}

\item{census_tract_to}{Destination Census Tract (string)}

\item{mode_of_transport}{Mode of transport (string) "subway" or "drive" or "walk"}

\item{key}{Google Maps API Key}
}
\value{
data.frame with columns "distance" (in meters) and "duration" (in seconds)
}
\description{
Travel Distance
}
