% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borough_sf_obj.R
\docType{data}
\name{borough_sf_obj}
\alias{borough_sf_obj}
\title{New York Borough SF Object}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 5 rows and 2 columns.
}
\usage{
borough_sf_obj
}
\value{
A sf object with column for \code{borough} (chr)
}
\description{
A simple features (sf) object of the five boroughs of New York City
}
\keyword{datasets}
