% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_symptomcount_variable.R
\name{calculate_symptomcount_variable}
\alias{calculate_symptomcount_variable}
\title{Calculate Symptom Count variable}
\usage{
calculate_symptomcount_variable(data)
}
\arguments{
\item{data}{Person data set}
}
\value{
data.frame with num_symptoms variable
}
\description{
Calculate Symptom Count variable
}
