% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_tract_sf_obj.R
\docType{data}
\name{census_tract_sf_obj}
\alias{census_tract_sf_obj}
\title{Census Tract SF Object}
\format{
sf
}
\usage{
census_tract_sf_obj
}
\description{
A simple features (SF) object of all census tracts in the United States
}
\keyword{datasets}
