% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_spatial_unit_ny.R
\name{convert_spatial_unit_ny}
\alias{convert_spatial_unit_ny}
\title{Convert Spatial Units in New York City}
\usage{
convert_spatial_unit_ny(
  census_tract = NULL,
  neighborhood_name = NULL,
  convert_to = "neighborhood"
)
}
\arguments{
\item{census_tract}{Vector (character) of input census tracts to convert to neighborhoods or boroughs}

\item{neighborhood_name}{Vector (character) of input neighborhoods to convert to boroughs}

\item{convert_to}{String showing how to convert census_tracts. Options are "neighborhood" (default) and "borough"}
}
\value{
Vector (character) of neighborhood or borrough names
}
\description{
Convert to parent spatial unit in hierarchy: borough > neighborhood > census_tract
}
