% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_spacial_neighbors.R
\name{get_census_spacial_neighbors}
\alias{get_census_spacial_neighbors}
\title{Get census spacial neighbors}
\usage{
get_census_spacial_neighbors(census_tracts)
}
\value{
List of neighbor sets
}
\description{
Find the neighboring census tracts for a vector of census tracts
}
