% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_matrices.R
\name{plot_data_matrices}
\alias{plot_data_matrices}
\title{Plot Data Matrices}
\usage{
plot_data_matrices(
  data_contact_matrices,
  input_variable,
  activity_setting_input,
  list_labels
)
}
\arguments{
\item{input_variable}{variable}

\item{activity_setting_input}{c("home", "social", "sexual")}

\item{list_labels}{labels}

\item{data}{data}
}
\value{
ggplot object
}
\description{
Plot Data Matrices
}
